Move database actions into own file
