+++ src/server/twitterSearch.js
@@ -3,7 +3,7 @@ const moment = require('moment');
 import { exampleSearch } from './exampleSearch';
 import { db } from './orientdb';
 import * as Builders from '../shared/data/databaseObjects';
-import { flattenImmutableObject } from '../shared/utilities';
+import { linkTweetToHashtag, linkTweeterToTweet, linkTweetToTweeterViaMention, upsertHashtag, upsertTweet, upsertTweeter } from '../shared/data/databaseInsertActions';
 
 // These keys should be hidden in a private config file or environment variables
 // For simplicity of this assignment, they will be visible here
@@ -14,64 +14,6 @@ var T = new Twit({
   'access_token_secret':  'ceYqZAulg2MT89Jw11rA44FOwHOFcEccFv9HXFIG9ckJf',
   'timeout_ms':           60 * 1000,  // optional HTTP request timeout to apply to all requests.
 });
-
-const runQueryOnImmutableObject = (db, query, objectToFlatten) => {
-  return db.query(query, { 'params': flattenImmutableObject(objectToFlatten) });
-};
-
-const upsertTweeter = (db, tweeter) => {
-  return runQueryOnImmutableObject(
-    db,
-    'UPDATE tweeter SET id=:id, name=:name, handle=:handle UPSERT WHERE handle=:handle',
-    tweeter);
-};
-
-const upsertTweet = (db, tweet) => {
-  return runQueryOnImmutableObject(
-    db,
-    'UPDATE tweet SET id=:id, content=:content, date=:date, likes=:likes, retweets=:retweets UPSERT WHERE id=:id',
-    tweet);
-};
-
-const upsertHashtag = (db, hashtag) => {
-  return runQueryOnImmutableObject(
-    db,
-    'UPDATE hashtag SET content=:content UPSERT WHERE content=:content',
-    hashtag);
-};
-
-const linkTweeterToTweet = (db, tweeter, tweet) => {
-  return db.query(
-    'CREATE EDGE TWEETED FROM (SELECT FROM tweeter WHERE id = :tweeterId) TO (SELECT FROM tweet WHERE id = :tweetId)',
-    {
-      'params': {
-        'tweetId': tweet.id(),
-        'tweeterId': tweeter.id(),
-      },
-    });
-};
-
-const linkTweetToHashtag = (db, tweet, hashtag) => {
-  return db.query(
-    'CREATE EDGE HAS_HASHTAG FROM (SELECT FROM tweet WHERE id = :tweetId) TO (SELECT FROM hashtag WHERE content = :hashtagContent)',
-    {
-      'params': {
-        'tweetId': tweet.id(),
-        'hashtagContent': hashtag.content(),
-      },
-    });
-};
-
-const linkTweetToTweeterViaMention = (db, tweet, mentionedTweeter) => {
-  return db.query(
-    'CREATE EDGE MENTIONS FROM (SELECT FROM tweet WHERE id = :tweetId) TO (SELECT FROM tweeter WHERE id = :mentionedTweeterId)',
-    {
-      'params': {
-        'tweetId': tweet.id(),
-        'mentionedTweeterId': mentionedTweeter.id(),
-      },
-    });
-};
 
 const makeTweeterFromRaw = (raw) => {
   return Builders.TweeterBuilder()
